refactor: 神クラス解消・コード重複排除
## 主な変更

### 新設ファイル
- `backend/api/resource_resolver.py`
  parse_model_id / require_character / require_preset / require_model_config を一元化。
  6箇所に散在していた「名前 or UUID でルックアップ → なければ 404」を排除。

- `backend/core/time_awareness.py`
  TimeAwareness dataclass + compute_time_awareness() を新設。
  api/chat.py・openai/router.py・group_chat/service.py の3箇所にあった
  同一の時刻認識ロジックを集約。

- `backend/core/embedding_migration_service.py`
  api/ui.py の _migrate_embeddings() (50行) を core サービスとして独立。
  APIレイヤーからビジネスロジックを分離。

- `backend/api/chat_images.py` / `backend/api/chat_drifts.py`
  api/chat.py (533行) から画像 API と SELF_DRIFT API を分割。
  セッション・メッセージ・画像・DRIFTの責務を別ファイルへ。

- `backend/core/memory/stores/` (7ファイル)
  SQLiteStore (935行) をドメイン別 Mixin に分割:
  SettingsStoreMixin / CharacterStoreMixin / MemoryStoreMixin /
  DigestStoreMixin / PresetStoreMixin / ChatStoreMixin / DriftStoreMixin。
  SQLiteStore はファサードとして多重継承し後方互換を維持。

### 変更ファイル
- `backend/core/chat/service.py`
  execute() と execute_stream() で約60%重複していた前処理を
  _prepare_context() コルーチンに抽出。デバッグログも _log_debug() に集約。

- `backend/api/utils.py`
  fmt_dt() と char_to_dict() を追加。21箇所の `dt.isoformat() if dt else None`
  を fmt_dt() に統一。characters.py の _char_to_dict() も utils に移動。
