+++ backend/api/ui.py
@@ -1,6 +1,5 @@
 """Settings UI routes using Jinja2 templates."""
 
-import asyncio
 import base64
 import uuid
 from typing import Optional
@@ -9,8 +8,7 @@ from fastapi import APIRouter, Form, Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 from fastapi.templating import Jinja2Templates
 
-from ..core.memory.chroma_store import ChromaStore
-from ..core.memory.manager import MemoryManager
+from ..core.embedding_migration_service import migrate_embeddings
 from ..core.providers.registry import (
     PROVIDER_LABELS,
     PROVIDER_ORDER,
@@ -48,70 +46,6 @@ def get_templates() -> Jinja2Templates:
     if templates is None:
         raise RuntimeError("Templates not initialized")
     return templates
-
-
-async def _migrate_embeddings(
-    app,
-    new_provider: str,
-    new_model: str,
-    new_api_key: str,
-) -> None:
-    """embeddingモデル変更時に全キャラクターの記憶を新モデルで再インデックスする。
-
-    SQLiteから全アクティブ記憶のテキストを読み出し、新しいembedding functionで
-    ChromaDBに再登録する。旧コレクションはdrop後に新モデルで再作成される。
-    app.state.chroma / memory_manager / chat_service も新しいインスタンスに差し替える。
-
-    Args:
-        app: FastAPIアプリケーション（app.stateへのアクセスに使用）。
-        new_provider: 新しいembeddingプロバイダー（"default" / "google"）。
-        new_model: 新しいembeddingモデルID（空の場合はプロバイダーのデフォルト）。
-        new_api_key: 新しいAPIキー。
-    """
-    from ..core.chat.service import ChatService
-
-    sqlite = app.state.sqlite
-    old_chroma = app.state.chroma
-    chroma_db_path = app.state.chroma_db_path
-
-    def _do_migrate():
-        """同期的に全記憶を再インデックスする。スレッドプール上で実行する。"""
-        new_chroma = ChromaStore(
-            db_path=chroma_db_path,
-            embedding_provider=new_provider,
-            embedding_model=new_model,
-            api_key=new_api_key,
-        )
-        characters = sqlite.list_characters()
-        for char in characters:
-            # 旧コレクションを削除してから新モデルで再作成する
-            old_chroma.delete_all_memories(char.id)
-            memories = sqlite.get_all_active_memories(char.id)
-            for mem in memories:
-                new_chroma.add_memory(
-                    memory_id=mem.id,
-                    content=mem.content,
-                    character_id=char.id,
-                    metadata={
-                        "category": mem.memory_category,
-                        "contextual_importance": mem.contextual_importance,
-                        "semantic_importance": mem.semantic_importance,
-                        "identity_importance": mem.identity_importance,
-                        "user_importance": mem.user_importance,
-                    },
-                )
-        return new_chroma
-
-    new_chroma = await asyncio.to_thread(_do_migrate)
-
-    # app.stateを新しいインスタンスに差し替える
-    app.state.chroma = new_chroma
-    new_memory_manager = MemoryManager(sqlite=sqlite, chroma=new_chroma)
-    app.state.memory_manager = new_memory_manager
-    app.state.chat_service = ChatService(
-        memory_manager=new_memory_manager,
-        drift_manager=app.state.drift_manager,
-    )
 
 
 # --- Dashboard ---
@@ -422,7 +356,7 @@ async def save_settings(
         # APIキー保存後の最新値を使用する
         current_google_key = store.get_setting("google_api_key", "")
         try:
-            await _migrate_embeddings(
+            await migrate_embeddings(
                 request.app,
                 new_provider=embedding_provider,
                 new_model=embedding_model,
