+++ backend/main.py
@@ -12,7 +12,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 
 from .adapters.openai import router as openai_router
-from .api import characters, memories, chat as chat_module, group_chat as group_chat_module
+from .api import characters, memories, chat as chat_module, chat_images as chat_images_module, chat_drifts as chat_drifts_module, group_chat as group_chat_module
 from .api import ui as ui_module
 from .core.chat.service import ChatService
 from .core.memory.chroma_store import ChromaStore
@@ -143,6 +143,8 @@ app.include_router(characters.router)
 app.include_router(memories.router)
 app.include_router(ui_module.router)
 app.include_router(chat_module.router)
+app.include_router(chat_images_module.router)
+app.include_router(chat_drifts_module.router)
 app.include_router(group_chat_module.router)
